+++ Introspect/ViewExtensions.swift
@@ -83,9 +83,9 @@ extension View {
     /// Finds a `UIScrollView` from a `SwiftUI.ScrollView`, or `SwiftUI.ScrollView` child.
     public func introspectScrollView(customize: @escaping (UIScrollView) -> ()) -> some View {
         if #available(iOS 14.0, tvOS 14.0, macOS 11.0, *) {
-            return introspect(selector: TargetViewSelector.ancestorOrSiblingOfType, customize: customize)
+            return introspect(selector: TargetViewSelector.siblingOfTypeOrAncestor, customize: customize)
         } else {
-            return introspect(selector: TargetViewSelector.ancestorOrSiblingContaining, customize: customize)
+            return introspect(selector: TargetViewSelector.siblingContainingOrAncestor, customize: customize)
         }
     }
     
@@ -157,7 +157,11 @@ extension View {
 
     /// Finds a `NSScrollView` from a `SwiftUI.ScrollView`, or `SwiftUI.ScrollView` child.
     public func introspectScrollView(customize: @escaping (NSScrollView) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.ancestorOrSiblingContaining, customize: customize)
+        if #available(macOS 11.0, *) {
+            return introspect(selector: TargetViewSelector.siblingOfTypeOrAncestor, customize: customize)
+        } else {
+            return introspect(selector: TargetViewSelector.siblingContainingOrAncestor, customize: customize)
+        }
     }
     
     /// Finds a `NSTextField` from a `SwiftUI.TextField`