+++ IntrospectTests/UIKitTests.swift
@@ -139,8 +139,8 @@ private struct ListTestView: View {
 @available(iOS 13.0, tvOS 13.0, macOS 10.15.0, *)
 private struct ScrollTestView: View {
     
-    let spy1: () -> Void
-    let spy2: () -> Void
+    let spy1: (UIScrollView) -> Void
+    let spy2: (UIScrollView) -> Void
     
     var body: some View {
         HStack {
@@ -148,13 +148,38 @@ private struct ScrollTestView: View {
                 Text("Item 1")
             }
             .introspectScrollView { scrollView in
-                self.spy1()
+                self.spy1(scrollView)
             }
             ScrollView {
                 Text("Item 1")
                 .introspectScrollView { scrollView in
-                    self.spy2()
+                    self.spy2(scrollView)
                 }
+            }
+        }
+    }
+}
+
+@available(iOS 13.0, tvOS 13.0, macOS 10.15.0, *)
+private struct NestedScrollTestView: View {
+
+    let spy1: (UIScrollView) -> Void
+    let spy2: (UIScrollView) -> Void
+
+    var body: some View {
+        HStack {
+            ScrollView(showsIndicators: true) {
+                Text("Item 1")
+
+                ScrollView(showsIndicators: false) {
+                    Text("Item 1")
+                }
+                .introspectScrollView { scrollView in
+                    self.spy2(scrollView)
+                }
+            }
+            .introspectScrollView { scrollView in
+                self.spy1(scrollView)
             }
         }
     }
@@ -316,18 +341,60 @@ class UIKitTests: XCTestCase {
         wait(for: [expectation1, expectation2, cellExpectation1, cellExpectation2], timeout: TestUtils.Constants.timeout)
     }
     
-    func testScrollView() {
+    func testScrollView() throws {
         
         let expectation1 = XCTestExpectation()
         let expectation2 = XCTestExpectation()
+
+        var scrollView1: UIScrollView?
+        var scrollView2: UIScrollView?
+
         let view = ScrollTestView(
-            spy1: { expectation1.fulfill() },
-            spy2: { expectation2.fulfill() }
+            spy1: { scrollView in
+                scrollView1 = scrollView
+                expectation1.fulfill()
+            },
+            spy2: { scrollView in
+                scrollView2 = scrollView
+                expectation2.fulfill()
+            }
         )
         TestUtils.present(view: view)
         wait(for: [expectation1, expectation2], timeout: TestUtils.Constants.timeout)
+
+        let unwrappedScrollView1 = try XCTUnwrap(scrollView1)
+        let unwrappedScrollView2 = try XCTUnwrap(scrollView2)
+
+        XCTAssertNotEqual(unwrappedScrollView1, unwrappedScrollView2)
     }
-    
+
+    func testNestedScrollView() throws {
+
+        let expectation1 = XCTestExpectation()
+        let expectation2 = XCTestExpectation()
+
+        var scrollView1: UIScrollView?
+        var scrollView2: UIScrollView?
+
+        let view = NestedScrollTestView(
+            spy1: { scrollView in
+                scrollView1 = scrollView
+                expectation1.fulfill()
+            },
+            spy2: { scrollView in
+                scrollView2 = scrollView
+                expectation2.fulfill()
+            }
+        )
+        TestUtils.present(view: view)
+        wait(for: [expectation1, expectation2], timeout: TestUtils.Constants.timeout)
+
+        let unwrappedScrollView1 = try XCTUnwrap(scrollView1)
+        let unwrappedScrollView2 = try XCTUnwrap(scrollView2)
+
+        XCTAssertNotEqual(unwrappedScrollView1, unwrappedScrollView2)
+    }
+
     func testTextField() {
         
         let expectation = XCTestExpectation()